Change model architecture, add first controllers

+++ database/migrations/2025_08_10_140030_create_payments_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('payments', function (Blueprint $table) {
             $table->id();
             $table->foreignId('appointment_id')->constrained()->onDelete('cascade');
-            $table->decimal('amount', 10, 2);
+            $table->integer('amount');
             $table->tinyInteger('status')->default(1);
             $table->timestamps();
         });
